[Maps] move joins from LayerDescriptor to VectorLayerDescriptor (#112427)
* [Maps] move joins from LayerDescriptor to VectorLayerDescriptor

* clean up ISource

* export isVectorLayer

* tslint

* tslint for ml plugin

* tslint apm plugin
