+++ x-pack/plugins/maps/common/index.ts
@@ -12,6 +12,7 @@ export {
   FIELD_ORIGIN,
   INITIAL_LOCATION,
   LABEL_BORDER_SIZES,
+  LAYER_TYPE,
   MAP_SAVED_OBJECT_TYPE,
   SOURCE_TYPES,
   STYLE_TYPE,
