+++ x-pack/plugins/maps/public/classes/sources/source.ts
@@ -53,8 +53,6 @@ export interface ISource {
   isESSource(): boolean;
   renderSourceSettingsEditor(sourceEditorArgs: SourceEditorArgs): ReactElement<any> | null;
   supportsFitToBounds(): Promise<boolean>;
-  showJoinEditor(): boolean;
-  getJoinsDisabledReason(): string | null;
   cloneDescriptor(): AbstractSourceDescriptor;
   getFieldNames(): string[];
   getApplyGlobalQuery(): boolean;
@@ -153,14 +151,6 @@ export class AbstractSource implements ISource {
 
   getGeoGridPrecision(zoom: number): number {
     return 0;
-  }
-
-  showJoinEditor(): boolean {
-    return false;
-  }
-
-  getJoinsDisabledReason(): string | null {
-    return null;
   }
 
   isESSource(): boolean {
