+++ x-pack/plugins/maps/public/connected_components/edit_layer_panel/edit_layer_panel.test.tsx
@@ -64,6 +64,9 @@ const mockLayer = {
   showJoinEditor: () => {
     return true;
   },
+  canShowTooltip: () => {
+    return true;
+  },
   supportsElasticsearchFilters: () => {
     return false;
   },
