+++ x-pack/plugins/maps/public/connected_components/edit_layer_panel/index.ts
@@ -13,11 +13,18 @@ import { LAYER_TYPE } from '../../../common/constants';
 import { getSelectedLayer } from '../../selectors/map_selectors';
 import { updateSourceProp } from '../../actions';
 import { MapStoreState } from '../../reducers/store';
+import { isVectorLayer, IVectorLayer } from '../../classes/layers/vector_layer';
 
 function mapStateToProps(state: MapStoreState) {
   const selectedLayer = getSelectedLayer(state);
+  let key = 'none';
+  if (selectedLayer) {
+    key = isVectorLayer(selectedLayer)
+      ? `${selectedLayer.getId()}${(selectedLayer as IVectorLayer).showJoinEditor()}`
+      : selectedLayer.getId();
+  }
   return {
-    key: selectedLayer ? `${selectedLayer.getId()}${selectedLayer.showJoinEditor()}` : '',
+    key,
     selectedLayer,
   };
 }
