+++ x-pack/plugins/ml/public/application/explorer/anomalies_map.tsx
@@ -18,6 +18,7 @@ import {
 } from '@elastic/eui';
 import {
   FIELD_ORIGIN,
+  LAYER_TYPE,
   SOURCE_TYPES,
   STYLE_TYPE,
   COLOR_MAP_TYPE,
@@ -125,7 +126,7 @@ export const getChoroplethAnomaliesLayer = (
       isTimeAware: true,
     },
     visible: false,
-    type: 'VECTOR',
+    type: LAYER_TYPE.VECTOR,
   };
 };
